Items Equip
The inventory system is now embedded into the main game scene.  Clicking on an item in the closet will equip it, putting a green check mark on the item box and placing it on the player cat.  Clicking another item will equip that one and unequip the previous one.  Clicking the same item will unequip it, leaving no item for that slot (head or body).